Moving remaining vis functions
