+++ pyreal/visualize/__init__.py
@@ -1,3 +1,16 @@
-from pyreal.visualize.feature_based_viz import plot_top_contributors, swarm_plot
+from pyreal.visualize.feature_based_vis import plot_top_contributors, swarm_plot
+from pyreal.visualize.time_series_vis import (
+    plot_time_series_explanation,
+    plot_shapelet,
+    plot_timeseries_saliency,
+)
+from pyreal.visualize.tree_vis import plot_tree_explanation
 
-__all__ = ["plot_top_contributors", "swarm_plot"]
+__all__ = [
+    "plot_top_contributors",
+    "swarm_plot",
+    "plot_time_series_explanation",
+    "plot_shapelet",
+    "plot_timeseries_saliency",
+    "plot_tree_explanation",
+]
